+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Support for G'MIC filters that produce multiple output images.
+
 ### Changed
 
 * The host application name is now set to `paintdotnet`.
 * Improved the exception documentation for `PdnGmicSharp`.
+* Renamed the `Output` property to `OutputImages` and changed the return type to `IReadOnlyList<PdnGmicBitmap>`.
 
 ## v0.1.0